Extract code for creating the transform stream into a separate module / file.
