+++ backend/utils/httpHeaders.js
@@ -1,5 +1,5 @@
 function setStreamResponseHeaders(res) {
-   return res.setHeader('Content-Type', 'application/octet-stream');
+    return res.setHeader('Content-Type', 'application/octet-stream');
 }
 
 module.exports = {
